Issue #1197: Added Support link in help menu and expired license warning
- Renamed "Docs" button to "Help" button after adding to it the support contact button.
- Renamed the component `DocsMenu` to `HelpMenu`.
- Tested on the UI.
